requirements.txt
pandas>=2.0
numpy>=1.24
python-dateutil>=2.8
PyYAML>=6.0
tabulate>=0.9
pytz>=2023.3
tzdata>=2023.3
